feat: moved `FollowerEvent` into event package
